app/requirements.txt
Flask==2.2.5
gunicorn==23.0.0
python-pptx==0.6.21
openpyxl==3.1.2
reportlab==4.0.9
